split out renderer related components
